fix reconcilation when pods was deleted
When pods are deleted then podRunnerPairs should be reloaded, so it's
better to requeue reconcilation immediately.

When new pods are scale up then it takes some time usually a few seconds
to sync between pods and runner API and so there is no need to wait
whole period (default 1m) to sync.

Signed-off-by: Lukas Hybner <git@hybner.net>